Week

+++ src/main/java/com/example/demo/mentoring/MentorPostService.java
@@ -1,7 +1,7 @@
 package com.example.demo.mentoring;
 
 import com.example.demo.config.errors.exception.Exception401;
-import com.example.demo.config.errors.exception.Exception500;
+import com.example.demo.config.errors.exception.Exception403;
 import com.example.demo.config.errors.exception.Exception404;
 import com.example.demo.mentoring.contact.ContactJPARepository;
 import com.example.demo.mentoring.contact.NotConnectedRegisterUser;
@@ -16,7 +16,10 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.time.Duration;
+import java.time.LocalDateTime;
 import java.util.List;
+import java.util.Locale;
 import java.util.stream.Collectors;
 
 @Transactional
@@ -29,67 +32,41 @@ public class MentorPostService {
 
     //mentorPost생성
     @Transactional
-    public void createMentorPost(MentorPostRequest.CreateDTO createDTO, User writer) {
-        if ( writer.getRole() == Role.MENTEE ) {
-            throw new Exception401("해당 사용자는 멘티입니다.");
+    public MentorPostResponse.MentorPostIdDTO createMentorPost(MentorPostRequest.CreateMentorPostDTO createMentorPostDTO, User writer) {
+        isMentor(writer);
+
+        Pageable recentPostTime = PageRequest.of(0, 1);
+        Page<LocalDateTime> recentPostDate = mentorPostJPARepository.createdRecentPost(writer.getId(), recentPostTime);
+        if(recentPostDate != null && !recentPostDate.getContent().isEmpty())
+        {
+            Duration duration = Duration.between(recentPostDate.getContent().get(0), LocalDateTime.now());
+            if (duration.toMinutes() < 5) {
+                throw new Exception403("5분이내로 글을 작성 할 수 없습니다.");
+            }
         }
 
-        //글자수 확인
-        if(createDTO.getContent().length() > 300){
-            throw new Exception404("글자수가 300자를 넘어갑니다.");
-        }
+        MentorPost mentorPost = new MentorPost( writer, createMentorPostDTO.getTitle(), createMentorPostDTO.getContent());
 
-        MentorPost mentorPost = new MentorPost( writer, createDTO.getTitle(), createDTO.getContent());
+        mentorPostJPARepository.save(mentorPost);
 
-        try {
-            mentorPostJPARepository.save(mentorPost);
-        } catch (Exception e) {
-            throw new Exception500("unknown server error");
-        }
+        return new MentorPostResponse.MentorPostIdDTO(mentorPost);
     }
 
    /* 1. mentorPostList를 조회
     2. 각 List당 writer별 writerInterests를 조회
     3. MentorPostDTO 생성*/
-    public List<MentorPostResponse.MentorPostAllDTO> findAllMentorPost(MentorPostCategoryEnum searchCategory, String keyword, int page) {
+    public List<MentorPostResponse.MentorPostAllDTO> findAllMentorPost(String search, String keyword, int page) {
         Pageable pageable = PageRequest.of(page,5);
-        Page<MentorPost> pageContent = null;
-        
-        //검색별 pageContent 검색
-        if(searchCategory == MentorPostCategoryEnum.NULL)
-        {
-            pageContent = mentorPostJPARepository.findAll(pageable);
-        }
-        else if(searchCategory == MentorPostCategoryEnum.TITLE)
-        {
-            pageContent = mentorPostJPARepository.findAllByTitleKeyword("%" + keyword + "%", pageable);
-        }
-        else if(searchCategory == MentorPostCategoryEnum.WRITER)
-        {
-            pageContent = mentorPostJPARepository.findAllByWriterKeyword("%" + keyword + "%", pageable);
-        }
-        else if(searchCategory == MentorPostCategoryEnum.INTEREST)
-        {
-            pageContent = mentorPostJPARepository.findAllByInterestKeyword("%" + keyword + "%", pageable);
-        }
-        else
-        {
-            throw new Exception404("검색 분류가 잘못되었습니다.");
-        }
 
+        MentorSearchCategory mentorSearchCategory = MentorSearchCategory.valueOf(search.toUpperCase(Locale.ROOT));
 
-        if(pageContent.getTotalPages() == 0){
-            throw new Exception404("해당 글들이 존재하지 않습니다");
-        }
+        Page<MentorPost> result = mentorSearchCategory.execute(keyword, pageable, mentorPostJPARepository);
+        List<MentorPostResponse.MentorPostAllDTO> mentorPostAllDTOS = result.stream().map(mentorPost -> {
+            List<UserInterest> writerInterests = userInterestJPARepository.findAllById(mentorPost.getWriter().getId());
+            return new MentorPostResponse.MentorPostAllDTO(mentorPost, writerInterests);
+        }).collect(Collectors.toList());
 
-        //List<MentorPost> mentorPostList = mentorPostJPARepostiory.findAll();
-        List<MentorPostResponse.MentorPostAllDTO> mentorPostDTOList = pageContent.getContent().stream().map(
-                mentorPost -> {
-                    List<UserInterest> writerInterests = userInterestJPARepository.findAllById(mentorPost.getWriter().getId());
-                    return new MentorPostResponse.MentorPostAllDTO(mentorPost,writerInterests);
-                }
-        ).collect(Collectors.toList());
-        return mentorPostDTOList;
+        return mentorPostAllDTOS;
     }
 
     public MentorPostResponse.MentorPostDTO findMentorPost(int id){
@@ -108,55 +85,30 @@ public class MentorPostService {
                 .flatMap(mentee -> userInterestJPARepository.findAllById(mentee.getMenteeUser().getId()).stream())
                 .collect(Collectors.toList());
 
-        MentorPostResponse.MentorPostDTO mentorPostDTO = new MentorPostResponse.MentorPostDTO(mentorPost, mentorInterests, menteeList, menteeInterests);
-
-        return mentorPostDTO;
+        return new MentorPostResponse.MentorPostDTO(mentorPost, mentorInterests, menteeList, menteeInterests);
     }
     @Transactional
-    public void updateMentorPost(MentorPostRequest.CreateDTO createDTO, int id, User writer) {
+    public void updateMentorPost(MentorPostRequest.CreateMentorPostDTO createMentorPostDTO, int id, User writer) {
         isMentor(writer);
 
         MentorPost mentorPost = mentorPostJPARepository.findById(id).
                 orElseThrow(() -> new Exception404("해당 글이 존재하지 않습니다."));
 
-        //글자수 확인
-        if(createDTO.getContent().length() > 300){
-            throw new Exception404("글자수가 300자를 넘어갑니다.");
-        }
+        validateUserAccess(writer, mentorPost);
 
-        try {
-            mentorPost.update(createDTO.getTitle(), createDTO.getContent());
-        } catch (Exception e) {
-            throw new Exception500("unknown server error");
-        }
+        mentorPost.update(createMentorPostDTO.getTitle(), createMentorPostDTO.getContent());
+
     }
 
     public void deleteMentorPost(int id, User writer) {
-
         isMentor(writer);
 
-        try {
-            mentorPostJPARepository.deleteById(id);
-        } catch (Exception e) {
-            throw new Exception500("unknown server error");
-        }
-    }
+        MentorPost mentorPost = mentorPostJPARepository.findById(id).
+                orElseThrow(() -> new Exception404("해당 글이 존재하지 않습니다."));
 
-    //생성 시간까지 조회하는 test service 코드 입니다
-    public List<MentorPostResponse.MentorPostAllWithTimeStampDTO> findAllMentorPostWithTimeStamp() {
-        List<MentorPost> pageContent = mentorPostJPARepository.findAll();
-        //List<MentorPost> mentorPostList = mentorPostJPARepostiory.findAll();
-        List<MentorPostResponse.MentorPostAllWithTimeStampDTO> mentorPostDTOList = pageContent.stream().map(
-                mentorPost -> {
-                    List<UserInterest> writerInterests = userInterestJPARepository.findAllById(mentorPost.getWriter().getId());
-                    if(writerInterests.isEmpty()){
-                        throw new Exception404("해당 카테고리는 존재하지 않습니다");
-                    }
+        validateUserAccess(writer, mentorPost);
 
-                    return new MentorPostResponse.MentorPostAllWithTimeStampDTO(mentorPost,writerInterests);
-                }
-        ).collect(Collectors.toList());
-        return mentorPostDTOList;
+        mentorPostJPARepository.deleteById(id);
     }
 
     public void changeMentorPostStatus(MentorPostRequest.StateDTO stateDTO, int id, User writer) {
@@ -164,18 +116,23 @@ public class MentorPostService {
         isMentor(writer);
 
         MentorPost mentorPost = mentorPostJPARepository.findById(id)
-                .orElseThrow(() -> new Exception404("해당 글이 존재하지 않습니다."));;
+                .orElseThrow(() -> new Exception404("해당 글이 존재하지 않습니다."));
 
-        try {
-            mentorPost.changeStatus(stateDTO.getMentorPostStateEnum());
-        } catch (Exception e) {
-            throw new Exception500("unknown server error");
-        }
+        validateUserAccess(writer, mentorPost);
+
+        mentorPost.changeStatus(stateDTO.getMentorPostStateEnum());
     }
 
     private void isMentor(User writer) {
         if ( writer.getRole() == Role.MENTEE ) {
-            throw new Exception401("해당 사용자는 멘티입니다.");
+            throw new Exception401("권한이 없습니다");
+        }
+    }
+
+    private void validateUserAccess(User writer, MentorPost mentorPost) {
+        // 예외처리
+        if ( mentorPost.getWriter().getId() != writer.getId() ) {
+            throw new Exception401("권한이 없습니다.");
         }
     }
 }
